requirements.txt
alembic==1.0.0
asn1crypto==0.24.0
bleach==3.0.0
blinker==1.4
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.4.0
cryptography==2.3.1
dominate==2.3.1
Flask==1.0.2
Flask-HTTPAuth==3.2.4
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-Moment==0.6.0
Flask-PageDown==0.2.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
ForgeryPy==0.1
get==1.0.3
httpie==0.9.9
idna==2.7
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.0
post==1.0.2
public==2.0.1
pycparser==2.19
Pygments==2.2.0
PyJWT==1.7.1
PyMySQL==0.9.2
python-dateutil==2.7.3
python-editor==1.0.3
query-string==1.0.2
request==1.0.2
requests==2.20.0
six==1.11.0
SQLAlchemy==1.2.12
urllib3==1.24
visitor==0.1.3
webencodings==0.5.1
Werkzeug==0.14.1
WTForms==2.2.1
